// src/entities/payment/model/Payment.ts
export enum PaymentStatus {
  PENDING = 'PENDING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  CANCELLED = 'CANCELLED'
}
// ...
export interface PaymentCardInfo {
  type: 'CARD';
  paydate: string;
  amt: string;
  appno: string;
  cardcd: string;
  cardname: string;
  card_noinf: string;
  quota: string;
}

export interface PaymentBankInfo {
  type: 'BANK';
  paydate: string;
  amt: string;
  bankcode: string;
  bankname: string;
}

export interface PaymentVbankInfo {
  type: 'VBANK';
  amt: string;
  bankcode: string;
  bankname: string;
  account: string;
  account_name: string;
  limitdate: string;
}

export type PaymentInfo = PaymentCardInfo | PaymentBankInfo | PaymentVbankInfo;

export interface PaymentResponseItem {
  p_idx: string;
  title: string;
  sessions: string;
  amount: string;
  brand: {
    name: string;
    code: string;
  };
  branch: {
    name: string;
    code: string;
  };
}

export interface PaymentAmountInfo {
  total_amount: string;
  discount_amount: number;
  point_amount: string;
  payment_amount: string;
}

export interface PaymentPointInfo {
  used_point: string;
  remaining_point: string;
}

export interface CashReceiptInfo {
  type: string;
  identityNum: string;
  amount: string;
  serviceCharge: string;
  taxFreeAmount: string;
  totalAmount: string;
}

export interface PaymentResponse {
  resultCode: string;
  resultMessage: string;
  body: {
    orderid: string;
    p_idx: string[];
    pay_info: PaymentInfo;
    cahereceipt_info: CashReceiptInfo | null;
    mp_info: number[] | null;
    items: PaymentResponseItem[];
    amount_info: PaymentAmountInfo;
    point_info: PaymentPointInfo;
  };
}